Fixed problems with property loading (simplified relative paths)
