+++ src/.dockerignore
@@ -1,3 +1,6 @@
+learning_service/output
+learning_service/dataset
+
 **/*_venv
 
 **/.dvc/
