+++ src/learning_service/.gitignore
@@ -1,3 +1,6 @@
+output
+dataset
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
