+++ src/learning_service/get_data.py
@@ -5,18 +5,12 @@ import os
 import shutil
 
 from common.color_module import ColorsPrinter
-from learning_service.var_names import VarNames
-from learning_service.dir_util import get_directory_from_settings_or_default
+from learning_service.config import settings, VarNames
 
-DATASET_DIR = get_directory_from_settings_or_default(
-    VarNames.DATASET_FOR_TRAINING_DIR,
-    os.path.join(os.path.dirname(os.path.abspath(__file__)), "dataset")
-)
-DATA_PATH = get_directory_from_settings_or_default(
-    VarNames.DATA_DIR,
-    os.path.join(os.path.dirname(os.path.abspath(__file__)), "data")
-)
+DATASET_DIR = settings[VarNames.DATASET_FOR_TRAINING_DIR.value]
+DATA_PATH = settings[VarNames.DATA_DIR.value]
 RESOURCES_DATA_PATH = "learning_service/data"
+
 
 def copy_data():
     """
@@ -28,6 +22,7 @@ def copy_data():
         os.mkdir(DATASET_DIR)
     shutil.copytree(DATA_PATH, DATASET_DIR, dirs_exist_ok=True)
 
+
 def copy_data_from_resources():
     """
     Copies data from application resources folder and creates
@@ -38,5 +33,6 @@ def copy_data_from_resources():
         os.mkdir(DATASET_DIR)
     shutil.copytree(RESOURCES_DATA_PATH, DATASET_DIR, dirs_exist_ok=True)
 
-if __name__=='__main__':
+
+if __name__ == '__main__':
     copy_data()
